requirements.txt
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
idna==3.6
pillow==10.2.0
pypng==0.20220715.0
python-dotenv==1.0.1
qrcode==7.4.2
redis==5.0.1
reportlab==4.1.0
requests==2.31.0
six==1.16.0
spotipy==2.23.0
typing_extensions==4.9.0
urllib3==2.2.1
